+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/models/FormRecognizerAudience.java
@@ -12,13 +12,13 @@ import java.util.Collection;
 @Immutable
 public final class FormRecognizerAudience extends ExpandableStringEnum<FormRecognizerAudience> {
     /** Static value AZURE_RESOURCE_MANAGER_CHINA for FormRecognizerAudience. */
-    public static final FormRecognizerAudience AZURE_RESOURCE_MANAGER_CHINA = fromString("https://cognitiveservices.azure.cn");
+    public static final FormRecognizerAudience AZURE_CHINA = fromString("https://cognitiveservices.azure.cn");
 
     /** Static value AZURE_RESOURCE_MANAGER_GOVERNMENT for FormRecognizerAudience. */
-    public static final FormRecognizerAudience AZURE_RESOURCE_MANAGER_US_GOVERNMENT = fromString("https://cognitiveservices.azure.us");
+    public static final FormRecognizerAudience AZURE_GOVERNMENT = fromString("https://cognitiveservices.azure.us");
 
     /** Static value AZURE_RESOURCE_MANAGER_PUBLIC_CLOUD for FormRecognizerAudience. */
-    public static final FormRecognizerAudience AZURE_RESOURCE_MANAGER_PUBLIC_CLOUD = fromString("https://cognitiveservices.azure.com");
+    public static final FormRecognizerAudience AZURE_PUBLIC_CLOUD = fromString("https://cognitiveservices.azure.com");
 
     /**
      * Creates or finds a FormRecognizerAudience from its string representation.