+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -34,6 +34,7 @@ and `customDocumentModelLimit` on model `ResourceDetails`.
   - `beginComposeModel` to `beginComposeDocumentModel`
 - Renamed property `DocumentFieldType.INTEGER` to `DocumentFieldType.LONG`
 - Renamed method `getValueInteger` to `getValueLong` on `DocumentField` model
+- Renamed properties on model `DocumentAnalysisAudience`
 
 ### Other Changes